pyproject.toml
[project]
name = "mobile-shopping-chat-agent"
version = "0.1.0"
description = "AI-powered mobile phone shopping assistant backend"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.117.1",
    "llama-index>=0.14.2",
    "llama-index-llms-openai>=0.2.0",
    "pydantic[email]>=2.11.9",
    "pydantic-settings>=2.10.0",
    "uvicorn>=0.37.0",
    "python-dotenv>=1.0.0",
    "pymongo>=4.15.1",
]
